Retrieves fund historical market data from Bloomberg.

Provided with a set of Bloomberg fund tickers and a time period, queries Bloomberg for the corresponding fund historical market data.

BBG_fund_market(tickers, start, end, verbose = T, ...)

Arguments

tickers
a chatacter vector. Specifies the Bloomberg fund tickers to query data for.

start
a scalar character vector. Specifies the starting date for the query in the following format: 'yyyy-mm-dd'.

end
a scalar character vector. Specifies the end date for the query in the following format: 'yyyy-mm-dd'.

verbose
a logical scalar vector. Should progression messages be printed? Defaults to TRUE.

...
optional parameters to pass to the bdh function from the Rblpapi package used for the query (options parameter).

Value

An S4 object of class FundMarket.

See also

Examples

if (FALSE) {

    BBG_fund_market(tickers = c("SPY US Equity", "IVV US Equity"),
      start = "2000-01-01", end = as.character(Sys.Date()))

  }